Metronomeクラスを作成、Cubeも追加
