+++ Assets/Scripts/Plain.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 85446a8cc2800a24597759c1a36911af
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
