+++ Assets/Scripts/Plain/Metronome.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 475dbb6563f69584984ad6e14a8e3b60
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
